+++ tests/Regressors/KNNRegressorTest.php
@@ -29,14 +29,14 @@ class KNNRegressorTest extends TestCase
      *
      * @var int
      */
-    protected const TRAIN_SIZE = 200;
+    protected const TRAIN_SIZE = 512;
 
     /**
      * The number of samples in the validation set.
      *
      * @var int
      */
-    protected const TEST_SIZE = 20;
+    protected const TEST_SIZE = 128;
 
     /**
      * The minimum validation score required to pass the test.
@@ -72,9 +72,9 @@ class KNNRegressorTest extends TestCase
      */
     protected function setUp() : void
     {
-        $this->generator = new HalfMoon(4.0, -7.0, 1.0, 90, 0.02);
+        $this->generator = new HalfMoon(4.0, -7.0, 1.0, 90, 0.25);
 
-        $this->estimator = new KNNRegressor(3, true, new Minkowski(3.0));
+        $this->estimator = new KNNRegressor(5, true, new Minkowski(3.0));
 
         $this->metric = new RSquared();
 
@@ -134,7 +134,7 @@ class KNNRegressorTest extends TestCase
     public function params() : void
     {
         $expected = [
-            'k' => 3,
+            'k' => 5,
             'weighted' => true,
             'kernel' => new Minkowski(3.0),
         ];